finished unsigned long long to qint64

+++ exportdialog.cpp
@@ -4,7 +4,7 @@
 // Copyright 2015 Pasquale J. Rinaldi, Jr.
 // Distrubted under the terms of the GNU General Public License version 2
 
-ExportDialog::ExportDialog(QWidget *parent, unsigned long long curcheckcount, unsigned long long curlistcount) :
+ExportDialog::ExportDialog(QWidget *parent, qint64 curcheckcount, qint64 curlistcount) :
     QDialog(parent),
     ui(new Ui::ExportDialog)
 {
